+++ spec/lips.spec.js
@@ -2,14 +2,6 @@
 
 var lips = require('../dist/lips');
 
-/* TODO
- *     test quasiquote
- *  (define (test)
- *       (let ((struct 'foo)
- *             (name 'bar))
- *         (print `(define (,(make-predicate name)  ,struct)))))
- */
-
 var {
     parse,
     tokenize,
@@ -182,6 +174,46 @@ describe('parser', function() {
                 )
             )
         );
+    });
+    function testQuasituote(input, literal) {
+        var tokens = tokenize(input);
+        var array = parse(tokens);
+        var q = new Symbol('quasiquote');
+        if (literal) {
+            q.literal = true;
+        }
+        var u = new Symbol('unquote-splicing');
+        if (literal) {
+            u.literal = true;
+        }
+        expect(array[0]).toEqual(
+            new Pair(
+                q,
+                new Pair(
+                    new Pair(
+                        new Symbol('list'),
+                        new Pair(
+                            new Pair(
+                                u,
+                                new Pair(
+                                    new Pair(
+                                        new Symbol('list'),
+                                        nil
+                                    ),
+                                    nil)
+                            ),
+                            nil)
+                    ),
+                    nil
+                )
+            )
+        );
+    }
+    it('should create quasiquote list from literals', function() {
+        testQuasituote('(quasiquote (list (unquote-splicing (list))))', true);
+    });
+    it('should use reader macros', function() {
+        testQuasituote('`(list ,@(list))');
     });
     it('should create AList', function() {
         var tokens = tokenize('((foo . 10) (bar . 20) (baz . 30))');
@@ -628,6 +660,51 @@ describe('lists', function() {
                 var input = str2list(code);
                 input.append(lips.nil);
                 expect(input).toEqual(str2list(code));
+            });
+        });
+    });
+});
+describe('cycles', function() {
+    it('should print simple cycles', function() {
+        var code = `(let ((x '(1 2 3))) (set-cdr! (cddr x) x) x)`;
+        return lips.exec(code + code + code).then(results => {
+            expect(results.length).toEqual(3);
+            results.forEach(result => {
+                expect(result.toString()).toEqual('(1 2 3 . #0#)');
+            });
+        });
+    });
+    it('should print cons withs cycles', function() {
+        var code = `(let ((x '(1 2 3))) (set-cdr! (cddr x) (cons x x)) x)`;
+        return lips.exec(code + code + code).then(results => {
+            expect(results.length).toEqual(3);
+            results.forEach(result => {
+                expect(result.toString()).toEqual('(1 2 3 #0# . #0#)');
+            });
+        });
+    });
+    it('should define list with two references', function() {
+        var code = `(define l (let ((x '(1 2 3))
+                                    (y '(1 2 3)))
+                                (set-cdr! (cddr x) x)
+                                (set-cdr! (cddr y) y)
+                                (cons x y)))
+                     l`;
+        return lips.exec(code).then(results => {
+            expect(results[0]).not.toBeDefined();
+            expect(results[1].toString()).toEqual('((1 2 3 . #0#) 1 2 3 . #1#)');
+        });
+    });
+    it('shoulde create double reference', function() {
+        var code = `(let ((x '(1 2 3))
+                          (y '(1 2 3)))
+                       (set-cdr! (cddr x) x)
+                       (set-cdr! (cddr y) y)
+                       (cons x y))`;
+        return lips.exec(code + code + code).then(results => {
+            expect(results.length).toEqual(3);
+            results.forEach(result => {
+                expect(result.toString()).toEqual('((1 2 3 . #0#) 1 2 3 . #1#)');
             });
         });
     });